db: Introduce driver table
Instead of the old tbl_user we rename this to driver,
since it is a table for having drivers int.
This table is then made compatible with auth_user, the django
provided auth table.
This allows us to later use our own custom table and model,
for easy creation of users with all the required data.