db: Generate models from database
With:
`docker-compose run web python manage.py inspect_transport_db`

I have tried several round of using AI to to improve on the
inspect command, but they all fail, and I think it is too time
consuming for me to do it.
I'll just go with this generated model, and modify from here by hand
(or with AI).

The AI is pretty good at modifying concrete code, but to modify Meta
code, i.e. code that generates/modifies code, seems to be outside of
it's current ability.