db: Make models and database match
As if created by Django.

First all the SQL was adjusted during the cleanup to use the
naming conventions of a hash from Django, so constraints are detected
by django.
Then the migration model was added to the initial migration, to match
the models of project, that again matches the SQL.
Then there is a separate migration that changes the models
to be managed by Django, and as they are already in sync, there
are no other changes.

To get the SQL required to make the changes I used
`docker-compose run web python manage.py makemigrations` to get
the migration model along with
`docker-compose run web python manage.py sqlmigrate transport 0002_...`
to get the sql changes.
These were incorporated back into the `cleanup.sql` code.

I then had ChatGPT 4o add ForeignKey handling.
I also had to manually add all the default values to all migration
models.

With this code the project gets started from the mysql loaded
by pgloader and does all changes required for working with the
models as native Django models going forward.